Optimize Unzip() by avoid a useless memory copy
`buf.Write(data)` create a slice and copy `data` into. Use
`bytes.NewBuffer(data)` to avoid this.
